refactor(api/applications): move document publish function to service
